flask-server/requirements.txt
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bidict==0.21.3
bleach==4.1.0
cffi==1.14.6
click==8.0.3
colorama==0.4.4
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.3
eventlet==0.32.0
Flask==2.0.2
Flask-PyMongo==2.3.0
Flask-SocketIO==5.1.1
gevent==21.8.0
greenlet==1.1.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pycparser==2.20
Pygments==2.10.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.5
python-engineio==4.2.1
python-socketio==5.4.0
pywin32==302
pywinpty==1.1.4
pyzmq==22.3.0
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
